refactor: WinnerFlag 계산 메서드 Player로 이동
